Delete MipUploadAPI parent class, derive child classes directly from UploadAPI Change in MipRNAUploadAPI that deletes a bunch of MipDNA functions that were included by accident
